New: This build of G Bars demonstrates the walking workflow end-to-end, INCLUDING data capture and marshalling into CSV.
Audio is fully synchronized, no stuttering, no overlaps.

The walking countdown is announced in 10-second intervals. It's more dramatic, and gives the (correct) impression the app is doing something in the mean time.

Pending the launch of the cloud(ish) server, the app provides an export button on the final page. This brings up the "Activity" sheet allowing many options for the export, including mail, but also inputs to other destinations, including storing in iCloud Drive, PDF, etc. Not all of these are useful.